+++ src/NetStandard/E.Standard.Platform/WebGISVersion.cs
@@ -6,7 +6,7 @@ public class WebGISVersion
 {
     public static Version Version
     {
-        get { return new Version(7, 25, 3202); }
+        get { return new Version(7, 25, 3401); }
     }
 
     public static string JsVersion